requirements.txt
paho-mqtt
requests
python-dotenv
pytz
pillow